+++ CHANGELOG.md
@@ -7,7 +7,8 @@
     - `build_update(type, data)` is dropped in favor of `deep_stringify(type => data)`.
 - Provide support for integration testing of bots in poller mode and non-Rails apps.
   __Breaking changes__:
-    - `telegram/bot/rspec/integration` is moved to `telegram/bot/rspec/integration/rails`.
+    - Requiring `telegram/bot/rspec/integration` is deprecated in favor of
+      `telegram/bot/rspec/integration/rails`.
     - `:telegram_bot` rspec tag is replaced with `telegram_bot: :rails`.
 
 # 0.13.1